# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Post API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/galaxies": {
            "get": {
                "tags": [
                    "Galaxies"
                ],
                "summary": "List all galaxies",
                "description": "Retrieve a list of galaxies with pagination",
                "operationId": "b2a628e9f49bf9c87a128584fa49a952",
                "responses": {
                    "200": {
                        "description": "List of galaxies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Galaxy"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Galaxies"
                ],
                "summary": "Create a new galaxy",
                "description": "Store a newly created galaxy resource",
                "operationId": "4c1d3bb39b1b7ab8fa9a7c332f3b8518",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Galaxy created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Galaxy"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/galaxies/{id}": {
            "get": {
                "tags": [
                    "Galaxies"
                ],
                "summary": "Get a specific galaxy",
                "description": "Retrieve a specific galaxy by ID",
                "operationId": "7af2bf3ef32f1be23286f4a2c2d74c91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the galaxy to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Galaxy details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Galaxy"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Galaxy not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Galaxies"
                ],
                "summary": "Update a specific galaxy",
                "description": "Update a specific galaxy resource",
                "operationId": "b28ee80691745270efdce00c82038788",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the galaxy to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Galaxy updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Galaxy"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Galaxy not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Galaxies"
                ],
                "summary": "Delete a specific galaxy",
                "description": "Delete a specific galaxy resource",
                "operationId": "3633da9f8639dbb09f4e53c62e293c7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the galaxy to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Galaxy deleted successfully"
                    },
                    "404": {
                        "description": "Galaxy not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Galaxy": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}